changed index to gallery compoentent and deleted meta data
