+++ pages/Gallery.tsx
@@ -1,0 +1,137 @@
+import type { NextPage } from 'next'
+import Head from 'next/head'
+import Image from 'next/image'
+import Link from 'next/link'
+import { useRouter } from 'next/router'
+import { useEffect } from 'react'
+import Modal from '../components/Modal'
+import cloudinary from '../utils/cloudinary'
+import getBase64ImageUrl from '../utils/generateBlurPlaceholder'
+import type { ImageProps } from '../utils/types'
+import { useLastViewedPhoto } from '../utils/useLastViewedPhoto'
+
+const Gallery: NextPage = ({ images }: { images: ImageProps[] }) => {
+  const router = useRouter()
+  const { photoId } = router.query
+  const [lastViewedPhoto, setLastViewedPhoto] = useLastViewedPhoto()
+
+  useEffect(() => {
+    // This effect keeps track of the last viewed photo in the modal to keep the index page in sync when the user navigates back
+    if (lastViewedPhoto && !photoId) {
+      document
+        .querySelector(`#photo-${lastViewedPhoto}`)
+        .scrollIntoView({ block: 'center' })
+
+      setLastViewedPhoto(null)
+    }
+  }, [photoId, lastViewedPhoto, setLastViewedPhoto])
+
+  return (
+    <>
+      <Head>
+        <meta
+          property="og:image"
+          content="https://nextjsconf-pics.vercel.app/og-image.png"
+        />
+        <meta
+          name="twitter:image"
+          content="https://nextjsconf-pics.vercel.app/og-image.png"
+        />
+      </Head>
+      <main className="mx-auto max-w-[1960px] p-4">
+        {photoId && (
+          <Modal
+            images={images}
+            onClose={() => {
+              setLastViewedPhoto(photoId)
+            }}
+          />
+        )}
+
+        <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 xl:grid-cols-3 2xl:grid-cols-4">
+          <div className="after:content relative col-span-1 row-span-3 flex flex-col items-center justify-center gap-4 overflow-hidden rounded-lg bg-white/10 px-6 pb-20 pt-0 text-center text-white shadow-highlight after:pointer-events-none after:absolute after:inset-0 after:rounded-lg after:shadow-highlight sm:col-span-2 lg:col-span-1 lg:row-span-2 lg:pt-0">
+            <h1 className="mt-8 mb-4 text-base font-bold uppercase tracking-widest">
+              Special Event Photos
+            </h1>
+            <p className="max-w-[40ch] text-white/75 sm:max-w-[32ch]">
+              Lorem ipsum dolor sit amet consectetur, adipisicing elit. Quos aliquam eaque cupiditate.
+            </p>
+            <a
+              className="pointer z-10 mt-6 rounded-lg border border-white bg-white px-3 py-2 text-sm font-semibold text-black transition hover:bg-white/10 hover:text-white md:mt-4"
+              href="/"
+              rel="noreferrer"
+            >
+              Back to Home
+            </a>
+          </div>
+          {images.map(({ id, public_id, format, blurDataUrl }) => (
+            <Link
+              key={id}
+              href={`/?photoId=${id}`}
+              as={`/p/${id}`}
+              id={`photo-${id}`}
+              shallow
+              className="after:content group relative cursor-zoom-in after:pointer-events-none after:absolute after:inset-0 after:rounded-lg after:shadow-highlight"
+            >
+              <Image
+                alt="Next.js Conf photo"
+                className="transform rounded-lg brightness-90 transition will-change-auto group-hover:brightness-110"
+                style={{ transform: 'translate3d(0, 0, 0)' }}
+                placeholder="blur"
+                blurDataURL={blurDataUrl}
+                src={`https://res.cloudinary.com/${process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME}/image/upload/c_scale,w_720/${public_id}.${format}`}
+                width={720}
+                height={480}
+                sizes="(max-width: 640px) 100vw,
+                  (max-width: 1280px) 50vw,
+                  (max-width: 1536px) 33vw,
+                  25vw"
+              />
+            </Link>
+          ))}
+        </div>
+      </main>
+      <footer className="p-6 text-center text-white/80 sm:p-12">
+        Noel Photography
+      </footer>
+    </>
+  )
+}
+
+export default Gallery
+
+export async function getStaticProps() {
+  const results = await cloudinary.v2.search
+    .expression(`folder:${process.env.CLOUDINARY_FOLDER}/*`)
+    .sort_by('public_id', 'desc')
+    .max_results(400)
+    .execute()
+  let reducedResults: ImageProps[] = []
+
+  let i = 0
+  for (let result of results.resources) {
+    reducedResults.push({
+      id: i,
+      height: result.height,
+      width: result.width,
+      public_id: result.public_id,
+      format: result.format,
+    })
+    i++
+  }
+
+  const blurImagePromises = results.resources.map((image: ImageProps) => {
+    return getBase64ImageUrl(image)
+  })
+  const imagesWithBlurDataUrls = await Promise.all(blurImagePromises)
+
+  for (let i = 0; i < reducedResults.length; i++) {
+    reducedResults[i].blurDataUrl = imagesWithBlurDataUrls[i]
+  }
+
+  return {
+    props: {
+      images: reducedResults,
+    },
+  }
+}
